// types/Fish.ts
export type Fish = {
    fishName: string;
    fishLength: string;
    fishWeight: string;
    fishingRodName: string;
    fishingRodLength: string;
    fishingRodTest: string;
    biteName: string;
    fishingLineType: string;
    photo: string;
    user?:string;
    fisherManName?:string;
    createdAt?:string;
    _id?:string
  }